Add option for sync or async write
When writing BEST_EFFORT samples in async mode, there is a higher chance
that some samples will be discarded and never put on the wire. The
write() method now gives the developer the option of performing the
operation either in asynchronous mode (when the caller provides a
callback) or synchronous mode (when a callback is not provided).